EMeetingStore: Remove deprecated EProxy usage.
EMeetingStore needs an ESourceRegistry property so it can obtain
an appropriate proxy ESource to serve as a GProxyResolver for the
SoupSession.

But frankly, this code is a mess and I can't be bothered right now.
The SoupSession will just use the desktop-wide proxy settings which
is fine for the vast majority of users.