EMailFormatter - Add new filter to remove META tags from message to avoid wrong encoding
We have to remove HTML meta tag with charset property set from
text/html parts of messages. Before we load message into web view Evolution
is transforming all messages to UTF-8. When WebKit2 loads the message that is
transformed by Evolution to UTF-8, but the message contains meta tag
that specifies the encoding WK2 prefers the encoding given by that meta
tag. So we have to remove this meta tag to show the message with right
encoding.