EEditorWidget: Fix smileys handling in composer
For every smiley in HTML mode (typed in text form or inserted from
menu) we create span wrapper that containes image and also text
representation. Afterwards when changing composer mode we are just
showing/hiding the unwanted one (in plain text mode we hide the
image representation and vice versa).