EEditorWidget: When getting plain text from composer wrap paragraphs in BODY clone
Don't do it in original BODY as we don't want to modify the body structure
for next calls. Before this change when we were sending HTML mail (we
are sending HTML and plain text version at same time) with
long lines the e_editor_widget_get_text_plain was issued before
e_editor_widget_get_text_html so even in HTML mail we sent the wrapped
text instead of unwrapped.