Bug 723377 - Add alternate zoom shortcuts
The zoom in shortcut (Ctrl+) can be confusing for some users because on
keyboards with a [+ =] key the '+' requires holding Shift.  To mitigate
the confusion, also accept (Ctrl=) for zoom in.

And while we're add it, also support Ctrl+ and Ctrl- on keypads, as well
as keyboards with dedicated Zoom In / Zoom Out keys.

(cherry picked from commit 8924b0bd5ef04c30ed3a43652e70e583407f7b85)