Add support for inline (CID) images
With this change all the inline images inserted into composer
(emoticons as well) are upon their insertion encoded to base64 and the
outcome is set to src attribute of the HTML image element. When sending
the message for all these images the CamelMimeParts are created and
attached to it. When saving message as draft no changes are made.  When
opening the message as new (and the message containes CID images) the
content from all the CamelMimeParts with images is encoded again to base64 and
saved. When the message is loaded the CID sources are replaced with
base64 data.The core functionality was moved from EMsgComposer to EEditorWidget
because right now it is too tight with DOM of the composer's web view.