EMailConfigAssistant: Hack the "Go Back" button after autoconfig.
After account autoconfiguration, you can click the "Go Back" button to
revise the account details.  That's been in place for some time now, but
it's not obvious from the button label.  "Go Back" implies going back to
the Identity page.

Temporarily rename the "Go Back" button on the Summary page to something
more accurate after a successful autoconfiguration, then reset it to its
original "Go Back" label.

Unfortunately the GtkAssistant API does not make this easy.  I had to
resort to crawling the assistant's internal child widgets and comparing
button labels to locate the right button.  So this hack may not be very
future-proof.