Bug 724909 - Highlight module hangs on large attachments
The previous code was writing the entire MIME part content to the
highlight utility's stdin pipe before reading the converted result.
With enough content, this caused the write operation to get stuck.
What's worse is this all happens synchronously in the UI thread.

Not sure exactly what was going on, but my hunch proved correct that
we need to simultaneously write to the stdin pipe and read from the
stdout pipe to avoid the deadlock.

Still not happy about this blocking the UI, but that would require
some major refactoring in libevolution-mail-formatter.