EWebView/EMailDisplay: Setup custom CSS styles to DOM right after document loads
The custom styles, those which are added on-the-fly directly to DOM, were
added only after whole document was completely loaded, which included also
sub-frames content, thus when these were slow to download for any reason,
the main frame content, like the top header bar, was shown with a default
style, which is white background and black text color. The WebKitWebView's
document-load-finished signal is emitted whenever the actual HTML content
is loaded, and the DOM is available as well, for each frame, including
the main frame of the WebView, thus updating/setting custom CSS styles
in this signal handler makes the loaded frames look as expected, without
waiting for other frames in the document to finish their load.