Contact photos not always shown in a contact preview
Locally stored photos were shown as a "missing image" instead of the actual
photo for cases when the URL contained special characters which required to be
encoded in the URL. This effectively updates commit 255e0529040baae67e3d70f1030ce248fd61856d,
to check WebKit version for the unescape, because WebKit 2.2.5 (which I currently use)
doesn't re-escape the URIs. Maybe earlier versions of the WebKit do the same.