Make 'Copy Link Location' copy into both primary and selection clipboards
It used to copy the link into both buffers, same as the 'Copy Email Address"
does, but it seems, after a move to WebKit based message preview, this
changed and the link had been copied only to the selection clipboard.
Make the function consistent and copy the link into to both clipboards.