Leftover "Generating message list..." info in a message-list
In some cases, like on errors or cancel, there could be left
an information in the message-list that the list is generating
its content, while it didn't do it in reality. This always resets
the message-list information in case the current message-list
generation finished and was the last in the queue.