Remove explicit adwaita-icon-theme dependency.
The last time we needed to explicitly depend on a particular g-i-t
version was when "default-avatar" was added, but that was nearly 4
years ago.

The new "adwaita-icon-theme" dependency breaks the build on anything
less than a bleeding-edge environment, so I think we'd be better off
just leaving the dependency implicit.

Alternatively we could still explicitly check for gnome-icon-theme as
a fallback for adwaita-icon-theme, but meh... it's not worth it IMO.