Custom mail account order reset on an account add or remove
It was quite discouraging to change order of mail accounts when it
reset to default sort order when a new account had been added or
an old removed. Rather than reset the order, the Evolution should
adapt the account count change properly, same as sort newly added
accounts by the default sort function.