Bug 731758 - [webkit-composer] Formatting/fonts not reset on paste into plain text
The cause of this bug was that WebKitWebView is not emitting signal
paste-clipboard for Shift + Insert (neither for Ctrl + Insert for copy
or Shift + Delete for cut) so we didn't get a change to process the
content of clipboard. As a workaround we are emitting paste-clipboard signal
when we trigger these shortcuts.