EHTMLEditorView - Improve the writing into quoted content
After this change when we start writing into quoted content it will
automatically rewrap the paragraph and reinserts the quote marks if the
block format is set to Normal. Also when the caret is in the beginning
of the quoted line and BackSpace is pressed the quoted content is
tooked out from the quote and inserted into body.