EHTMLEditorView - Replace 2+ spaces with non-breaking spaces and create anchors for links when processing data from convertor
When converting text we want to preserve spaces from original text. For this we
have to use the non-breaking spaces for it as two and more normal spaces is WebKit
replacing with just one. But we want the text from convertor to be wrapped
(non-breaking space is preventing from it) so we will replace 2+ spaces
with non-breaking spaces and leave the space of length one as it is.

Also when we are going through the text from the convertor we will
surround the links with anchor element to make them clickable.