Reply to selection in text-highlight part loses white-spaces
text-highlight module encloses the generated text into <pre/> tags,
thus the preview panel shows white-spaces correctly, but once
a reply is done to a selected text the returned HTML portion had
missing <pre/> tags, which led to lose of the white-spaces in
the generated content (in UI), while the white-spaces were there
in the raw HTML code. Enclosing selected HTML portion into <pre/>
tags in case the original text is also enclosed it them fixes it.