When setting some content into the composer body we have to specify if it is a HTML content
Code paths for HTML content and plain text content are different. When
we will try to load plain text content as a HTML content we will lose
the new line characters. In the opposite case when HTML content is load
as a plain text at least the whole structure is lost.