Always use selected color for selected region in EDayView
It used focused or unfocused selected color, but the EDayView itself
is not focused at all, it's main_item->parent.canvas instead. Instead
of trying to hunt for this (and updates when the focused widget changes)
just use the same selected color unconditionally, just the way the other
views do it.