Bug 733954 - Enter doesn't delete selection with cited text
The problem was that we were manipulating with the selection before the
WebKit Editor call (and we hit again problems with selections in
WebKit). So as a fix we reworked how the new line is inserted when the
enter is pressed on empty line in quoted content (and completely avoid
the WebKit Editor call there).