EHTMLEditor - Change the way how the Indent/Unindent and Alignment works
Before this change the indent was done with BLOCKQUOTE element. But the
BLOCKQUOTE elements have marings around them set by the HTML engine.
For composer we can override them, but when someone will receive that
HTML message he will have unwanted margin around them. To solve this we
will use the DIV elements instead.
Also for the alignment we switched away from WebKit Editor, but we are
doing it ourselves.