Bug 736476 - Space lost on the end of line
Again this had something to do with WebKit behavior regarding to the
DOM selection. The thing is that when the selection is saved and
afterwards restored on the same position it is not restored on the same
position. So we hav to determine if we are writing into the quoted text
differently without saving the selection.