Improve and fix various dialogs.
In EHTMLEditorCellDialog and EHTMLEditorTableDialog encode the
background images for elements into the base64 format to prevent
from deleting them when sending the HTML message. Also set the default
background color in these dialogs to transparent. In dialogs previously
mentioned and as well as in EHTMLEditorPageDialog add the button to
remove the previously set background image.