Add metainfo.xml files for bogofilter and spamassassin plugins
These are useful for GNOME Software, to show the plugins as an addon
for the Evolution. Packagers can decide what to do with the file,
either they can add it into the respective subpackage or they can
just ignore (remove) the files, if they do not build special packages
for these plugins.