[Alarm Notify] Avoid UI freeze on an alarm dismiss
There was called a synchronous version of the ECalClient function
in the main thread, which caused the UI freeze when the backend
(or calendar factory) had been busy with something. This was fine
some time ago, when this function was invoked in a dedicated thread,
but that's not a case for some time. Using async version of the function
avoids the UI freeze.