EHTMLEditorView - Try to preserve the text blocks when processing the plain text content
Previously all the lines from plain text content were inserted into the
composer each as separate block. With this patch we added the logic to
preserve the blocks if possible or insert the line as separate block
when it is not (this happens when the line starts with space or some
special characters).

Also fix various bugs in the parse_html_into_paragrahs function.