EHTMLEditorView - Don't use separate WebView for converting the HTML message to plain text, but do it in situ
One reason for it is the WebKit2 port (the separation of UIProcess and
WebProcess -> this would be more complicated) and the another one was
performance. On complicated HTML test message the "Edit as new" action
with "Lose Formatting" answer could lead to nearly 30 second of
processing various things. Now the gap there is ~2 seconds.