EHTMLEditorView - Fix logic that prevents the text blocks when processing the content into the composer
End block if the current line has less than 62 characters and the
length of the first word on the next line (if presented) plus the
length of the current line is less than 72 characters (that means that
the line was wrapped as there was not enough space left for that word).
Also end block if the line has more than 72 characters.