Revert "[EMailBackend] Provide cancellable and show in UI prepare for a quit"
This reverts commit 7c918831a8aac463ffe3cdd7e39b36ad09eec34a.

It turned out that the Evolution won't quit when the quit was initiated
with a "busy" EShellBackend. In that case the EShellBackend itself
references the activity and lets it alive until the activity is completed.
The problem is that the activity is completed only after it is freed,
thus there happens a circular dependency on the reference count of
that activity.