Revert "[EMailBackend] Provide cancellable and show in UI prepare for a quit"
This reverts commit 7c1aba6df6695e22900188c6b8e01549f391b79d.

It turned out that the Evolution won't quit when the quit was initiated
with a "busy" EShellBackend. In that case the EShellBackend itself
references the activity and lets it alive until the activity is completed.
The problem is that the activity is completed only after it is freed,
thus there happens a circular dependency on the reference count of
that activity.