Odd cross line when drawing an ECalendar with multiple columns
Multiple columns mean multiple months on a row, which can be seen
for example in a Recurrence definition dialog when editing an event.
The line was meant as a divider between months, but it had flipped
x and y destination coordinates. After fixing that the line looked
rather distracting, thus I removed it completely.