Rework how the selection inside the qouted content is deleted.
The thing is that normally the quote marks are not selectable by user. But
this caused a lof of problems for WebKit when removing the selection. This will
avoid it as when the delete or backspace key is pressed we will make the quote
marks user selectable so they will act as any other text. On HTML keyup event
callback we will make them again non-selectable.