Special-case Google 'caldav' calendars to use 'google' editor
When a user created a Google calendar, in a Google editor, then a CalDAV
backend was created. Opening Properties of the Google calendar opened
the CalDAV editor instead, because it was backend-driven which editor
will be opened. This special-case addition makes sure that Google
calendar properties will open Google editor, the same which was used
to create the calendar.