Bug 744849 - Insert->Emoticon deletes text before cursor
The problem was that it was expecting that the newly inserted smiley
was written (and not inserted through the picker). So we have to
remember if the smiley was written or not. Another problem was that we
were checking just the first character of the written smiley
(typically just colon). I also rewrote the string manipulations around
the smileys into DOM operations.