Bug 743404 - Status column of message list exposed poorly via atk
Allow the atk implementation for ECellToggle to expose names for
states, since an ECellToggle can potentially show many states,
so setting ATK_STATE_CHECKABLE isn't always sufficient. Also
expose the corresponding enum of the icon through AtkValue, in case
an AT wants to perform a custom action for a particular value. Use
this functionality for the "status" cells in the message list.