Bug 745551 - Pasting text from gnome-terminal with concatenated whitespace and tabulators, some whitespace occurrences get removed
The problem was the the code expected that the matched pattern will be
just the spaces or tabulators nor both of them at the same time. As
a fix we need to process the matched pattern characters one by one.

Also when deciding whether to preserve the block (line) we need to
operate with a string that already has the tabulators replaced.