Bug 745899 - Shift+Tab in message body doesn't move focus to headers
The behavior right now is. When tabulator is pressed in view (outside
of table) tabulator is inserted. When tab is pressed inside a table,
caret is moved to next cell in the table. If we are in the last cell of
table, caret is moved to the block after table. When Shift+Tab is pressed
inside the view (outside of table) caret jumps to headers. When
Shift+Tab is pressed in table the caret is moved to previous cell. If
the caret is in first cell of table it is moved to the block that is
before the table. If there is no block before the table, focus is moved
to headers.