EHTMLEditorView - Introduce function that performs spell check just in current viewport
The e_html_editor_view_force_spell_check is heavy for most cases. On
some big mails composer can spent there even 10 seconds (on bigger
mails probably even more). With this commit we introduced the
e_html_editor_view_force_spell_check_in_viewport that really performs
the spell check on elements that are currently in viewport. The
downside of this approach is that when we scroll the viewport we the
newly visible elements won't be spell checked. As a solution we need to
perform the spell check when the scroll occurred (this information we
get from HTML scroll event). The problem was that the scrolling was
choppy as the e_html_editor_view_force_spell_check_in_viewport function
still need some time to perform the spell check thus we are performing
the spell check one second after last scroll event occurred.