EHTMLEditorSelection - Correctly save the selection markers in quoted content
Before this change the selection markers could be saved next to the
quote character and this could later prevent the selection to be
restored correctly. If the selection could be saved next to the quote
quote character we will save it before the first child of the block
that is quoted.