Message credits are put on the end of quoted content instead of on its beginning.
I received one message and on its end there was the paragraph element left opened:

	</p><p></body></html>

When we appended the credentials after it:

	</p><p></body></html><span class="-x-evo-to-body" data-credits><pre>credits</pre></span>

Now WebKit tried to do its best to parse, fix and load the content correctly that resulted in:

	<p><span class="-x-evo-to-body" data-credits></span></p><pre>credits</pre></body></html>

Clearly, we will fail when moving the credits to right place as we are
expecting the the credits are inner part of the span. So set the
credentials into the data-credits attribute (as there it will be
resistant against this behavior) and rework the part that's handling it.