Fails to find nested message's attachment bar in WebKitWebView
A runtime warning like this:

    evolution-mail-WARNING **: Failed to find parent <object>
    for '.message.mixed.1.mbox.0.rfc822.attachment-bar' - no ID set?

could be shown when a message in the preview had a message attachment,
which had also an attachment.

The reason for it was a use of the same variable name in the 'for' body
as in the parent block, which, even the element was found, didn't set
the parent block's variable, thus it behaved like nothing was found.