EHTMLEditorView - Setting the link color doesn't work
There were two issues. First one was that we were setting the link
color to the user style and it override the value that we later
set to link attribute on BODY element in Page dialog. The second issue
was that the if we changed the link attribute value, the color of existing
anchors in view sometimes didn't changed at all (probably because the
body is contenteditable). As a solution we are now changing the color
of links through the CSS style sheeet in HEAD as well.