Bug 747055 - Line full of equals signs halts selection of quoted text with shift + page down
The problem was that we were not splitting the long sequences of text (in this
case equals signs or even the long links) without a space that can't fit into
Normal block. If a long link is wrapped, the link will be removed and replaced
with the inner text of ANCHOR that will be wrapped later.