Composer - Extra new line added on the end of reply
When replying on message there is an extra empty line added on the end of the original
text. It gets there while parsing the original message as the trailing CRLF is not
removed. The proper fix would be to delete the trailing CRLF, but it is safer
to just mark it in editor to remove the trailing new line later.