Bug 747510 - Add composer option "Inherit theme colors in HTML mode"
With this change the EHTMLEditorView is using the colors from current theme.
Unless the newly introduced option "Inherit theme colors in HTML mode" is set
the colors are not sent in resulting HTML formatted message (white background
and black font are used).

With this patch the visited links (opened with Ctrl + Left click) are marked
as visited with the possibility of changing its color in Page properties dialog.