Bug 747996 - Cannot apply TT on multiple lines at once
When setting the i.e. bold format we are using the WebKit's build in editor
functions, but WebKit doesn't have the same for TT. The previous implementation
was very limited and non accurate. The previous implementation allowed i.e. monospaced
elements inside another monospaced elements or it didn't concatenate the adjacent
monospaced elements.

This changed also rewrites how the format for current selection is detected as the
previous implementation led to various false positives. Also it removes the unwanted
code duplication.