Bug 748178 - Eats CPU and memory when replying to calendar invite
The main cause was that composer went to busy loop when it was trying
to wrap a line with word that was longer that line length (usually 72
characters). Now we will force the split at line length to avoid this
situation.

Another thing was that when we wrapped the line it could happen that a
wrong text node will be chosen as a next node to split.