EHTMLEditorView - Implement undo and redo on text operations in table
The problem here is that we can't use the same solution for other
operations as the webkit_dom_document_caret_range_from_point will
return range that will just contain the focused TD element and not the
exact caret position. This is the WebKit limitation (the same applies
for TEXTAREA elements) so we have to save the whole TD element instead
of just saving what was added or removed.