EHTMLEditorView - Fix the 'Paste Quotation' action
The pasted and quoted text was wrongly inserted into another PRE element
that caused another parts of editor to misbehave (as it should be inserted
straight into BODY element). Also the pasted content was not preprocessed at
all before it was pasted and that also caused some undefined behaviour as well.