EHTMLEditorSelection - Fix text wrapping
In find_where_to_break_line we need to look at the current character
before we will try to break the line, otherwise when the last character
is space we would use the previous space as we didn't checked the current
one.

Also fix runtime critical warnings when we are on the end of the current
paragraph and we don't have any node to continue with.