EHTTPRequest - GFileInfo not valid when machine is under heavy load
When the machine is under heavy load and I'm moving between messages with
inline images (that are already downloaded and are in the cache) I get warnings
that GFileInfo object is not valid. I managed to get these warnings when the task
was already cancelled (I already moved to another message) but for the active task
as well. So try to access the GFileInfo object just when we know that it is valid.