Composer - Undo/Redo on operations with selection could be wrong
Previously, i.e. when we selected some text and pressed the Enter the
selected text is removed and later the new line is inserted. In some of
these situations the removal of the selected text was not saved at all that
later led to inconsistent undo/redo operations. Also the HISTORY_AND event
was introduced to add a possibility to distinguish the history events that
are supposed to be processed as one.