EHTMLEditorView - Undo/Redo around links is wrong
It could happen that when redoing the INPUT operations on link we could
obtain not the same link as originally. The same applies for undoing the INPUT
operations where we would just remove the characters from the link, but we will
leave the link not as plain text, but as valid anchor.