EHTMLEditorSelection - Text in quoted content is wrongly wrapped when contains tabulator
If we are replying to message we always wrap its content to width that is set
in Preferences - quotation level * 2 (length of "> "). The problem here is that
previously the wrapping code didn't know anything about the quotation level * 2
part so when it was wrapping a line with tabulators it was not counting with these
extra characters on the beginning of line thus it could happen that we would end
up with more characters on the line that was expected thus we would have another
extra line with text, but without quote character before it.