Bug 750615 - Using Ctrl+Z in plain text composer messes up content
The code that was there for undo/redo of block delete operation didn't handled
the quoted content as expected. The problem there was the mix of WebKit's editor
commands and our DOM operations and the fact that the quote characters are not
user selectable. In the end we need to correct what was actually removed to
correctly undo/redo it afterwards.