EHTMLEditorSelection - Alignment could be wrong
When determining alignment we have to look at element's classes not
just in case it is a list item, but in general as other elements
can have these classes set. Also when changing the block format save
the alignment before the change as after the change the value
in selection->priv->alignment will be reseted thus we wouldn't preserve
the alignment across format change.