EHTMLEditor - Improve lists formats
With this change the "Bulleted" list will have star character as a bullet
instead of a default HTML one. This applies for Plain Text mode, the HTML one
is left unchanged and the list bullets are given by WebKit.

With this change the lists are more compact as the indentation level is now set
to 3 characters (before 4) and the list level is set to 3 (from 8) as well. For
ordered lists (Numbered, Roman Numeral, Alphabetical) the initial indentation is
set to 6 as there won't be enough space for counters (especially for Roman numerals).

This also removes some definitions around lists alignments as they are not needed
anymore.