Editor - Width limits are wrong for lists when switching composer modes
If switching from Plain text mode to HTML mode and back with some lists in
message the width limits are correctly removed for HTML mode, but incorrectly
set for Plain text mode. Actually it is enough to set the width limit just to
the main blocks and not to sub-blocks as the width limit is inherited.

Also don't set an alignment for lists when setting width limits as the lists
are handled separately.