Composer - Leave copy&pasting inside the composer on WebKit
Let WebKit handle the copy&pasting of content inside the composer. This will
avoid some bugs like losing all the formatting (indentation, alignment) of the
content when copy&pasting it. When pasting content from outside of composer we
have to process and insert the content ourselves (to remove its formatting, parse
it to the composer's HTML structure and correctly insert it) as before this change.