EHTMLEditorView - Lines of text appended after the signature are sent on one line
If the user edited the signature or added more content after it, WebKit just
duplicated the DOM structure and left us with multiple PRE elements that don't
have the BR elements on their end. The content is rendered fine (every pre has
its own line), but when we try to get a plain text version of the signature we
will get the text from these PRE elements on one line. As a solution we need to
insert the BR elements on the end of earch PRE element (if not presented) to get
the correct text from signature.