Composer - User changed signature is lost when opening a draft or template
When opening a draft or a template with a signature we were setting the current
signature in "Signature chooser" to the one that was inside the message. The
problem was that the code replaced the body's signature with the one that is
saved on local storage. When the signature in the body was changed by the user
the changes were thrown away. After this change the signatures (the one in the
body and the one from local storage) are compared and if they are not the same
the signature in the body is left untouched and current signature in "Signature
chooser" is set to None (to stick with GtkHTML composer).