Bug 752997 - Crash when replying to the message
The problem is that WebKit1 (2.4.9) crashes when it is trying to move or remove
an anchor element that has an image element inside and this image element has
the CSS float property set in the style attribute. To workaround it we will rename
the style attribute and rename it back when we will send the message. It is
unfortunate that we can change the formatting with this, but this is definitely
better than crashing. This could be removed once Evolution switches to WebKit2
as the WebKit2 is unaffected (tested on 2.8.4).