EHTMLEditorView - Treat non-breaking space as a regular space
The problem is that before this change the same text could end wrapped differently
when written or when when it was pasted. When writing a text WebKit uses a
non-breaking space when pressing a space and then replacing it with regular space
if we continue with writing. Sometimes (after manipulating with selection) the
non-breaking space is not replaced with regular one thus the wrapping will fail
there. To avoid it we have to treat non-breaking space as a regular space.