Correct categories_icon_theme_hack() function
It adds an icon directory for one of the categories, which not always
works, especially when a translator in the evolution-data-server uses
a different category name translation, than the one from there. It is
better to traverse all categories and add all icon folders into the
search patch for the icon theme.