[ESpellEntry] Do not touch entry when its IM is in the preedit mode
When a user enters text with a preedit involved, the ESpellEntry
should not touch the attributes of the entry, because it breaks
the preedit buffer reset logic. It's not visible with the first word,
but entering the second without a delimiter shows garbage. The attribute
change also garbled the preedit highlighting in the entry.