EHTMLEditorSelection - Fix wrapping of a quoted content when editing a link inside it
Correctly traverse the DOM when looking for BR elements that needs to be removed and
when looking for the last BR element that will be used as a start point when wrapping.
Also we need to replace the hidden spaces just in that part of a paragraph that will
be wrapped and not in the whole paragraph.