EHTMLEditorSelection - Correctly wrap text nodes around selection markers
The word could be split by selection markers and in this case we would wrap it
wrongly as the other part of the word is in another text node after the marker.
As a solution temporary append the text of a node that is after the caret to the
node that is before the caret and remove this appended text after we wrap the text
that is before the caret.