EHTMLEditorView - Modifying the text of a smiley should convert the smiley to a text node
If deleting something in a smiley it won't be a smiley anymore (at least from
Evolution' POV), so remove all the elements that are hidden in the wrapper and leave
just the text. Also this ensures that when a smiley is recognized and we press the
BackSpace key we won't delete the UNICODE_HIDDEN_SPACE, but we will correctly delete
the last character of smiley.