Bug 756883 - Add an option for line wrapping in a quoted text
This patch adds a new "Wrap quoted text in replies" option under the
"Edit->Preferences->Composer Preferences->Replies and Forwards" section in
preferences. If set to FALSE the quoted text in replies will be parsed to
'Preformated' blocks and the won't be wrapped. The option is set to TRUE by
default, so the behavior won't change for current users.

Also rename various variables/functions around code that parses the HTML to
blocks to reflect that we are not parsing just to paragraphs.