Bug 757348 - Endless loop in webkit_dom_node_append_child(): assertion 'WEBKIT_DOM_IS_NODE(self)' failed
The main cause was that after modifying/deleting an original quoted content
there were some empty blockquote elements left and they caused the Evolution to
misbehave. It showed up that we are not removing the empty blockquotes after
deleting/modifying a quoted content, but only empty blocks inside them. Correct
the remove_empty_blocks to remove the empty blockquotes as well.

The loop took a place in the remove_wrapping_from_element function. To avoid it
return early if we pass an empty element to it and apply the same to
remove_quoting_from_element.