[EActivityBar] Do not remove timeout GSource twice in dispose
When the window holding the activity bar is closed, the widget itself
has run dispose. it's possible that the timeout source holds the last
reference to the widget, thus when the GSource is removed, it also
unreferences the widget, which calls the dispose again and it is
tried again to remove the same GSource. using a temporary variable
and setting the private variable to zero before the removal fixes
the critical warning which this double-remove causes.