Bug 757887 - Reproducible freeze when deleting quoted lines in email composer
The main problem was in the merge_siblings_if_necessary (also fixed the spelling
there) where we did not marked the already processed blockquotes, thus we were
starting from the beginning again. Also we can remove the -x-evo-main-cite id
from the blockquote (if presented) as it is not needed there anymore and it
could prevent the processing of that particular blockquote.

Also fix the undo and redo problems found with this particular message.