[EMailLabelListStore] Avoid circular dependency in a tag cache
The store has its own cache of tags for quicker lookup, but this
cache used GtkTreeRowReference-s, which reference the model, thus
it made circular dependency and the store was never freed. Using
place GtkTreeIter pointer in the cache works too, one only needs
to do more often manual updates of it.