[ECompEditor] Change order of an alert add and remove
The previous order of "add new, then remove old" was meant to workaround
bug in GtkInfoBar which can be left hidden due to the animation of it when
it contains the alert. It was forgotten that the alert bar avoids duplicates,
thus it could happen that the alert had been removed from the UI, because
it was already in the alert bar.

The previous commit workaround the GtkInfoBar issue differently, by disabling
the animation of it, which makes things much easier in the code.