Ask for a forced quit when a user tries to re-close the window
When the user calls Quit, the windows are disabled and the Evolution
prepares for quit. In case the quit takes longer than a minute, the user
is asked whether the Evolution should wait for the background tasks to
finish or quit. Sometimes the user wants to quit even earlier, thus let
it do it when it tries to close the window again.