Attachment's List View passes mouse events to the parent widget
That means that the parent widget, like EWebView, processes the event,
even when it was used by the GtkTreeView descendant, which results in
undesired behaviour, like a new mail message composer being opened
for one of the message recipients (click on an address in the To/CC
headers). The icon view doesn't do that, it reports the event as being
used, but let's make it the same for both views.

Reported downstream: https://bugzilla.redhat.com/show_bug.cgi?id=1285698