Bug 757939 - Undo failure after dragging text
The problem was that after we started to rely on WebKit while doing drag and drop
within the view we were not saving the history at all. This change implements
the history saving for drag and drop and fixes various problems related to drag
and drop as sometimes the content was copied instead of moved and sometimes the
dropped content (from outside of composer) was inserted twice.