Bug 749712 - Undo/Redo in composer doesn't work right
Implement the workaround on the Evolution side. The cause of the bug is that
WebKit is having problems determining the right line height for some fonts and
font sizes (the right and wrong value differ by 1). To fix this we will add an
extra one to the final top offset. This is safe to do even for fonts and font
sizes that don't behave badly as we will still get the right element as we use
fonts bigger than 1 pixel.